requirements.txt
absl-py==0.7.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
backcall==0.2.0
black==21.4b2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==1.6.0
configparser==5.1.0
cycler==0.11.0
Cython==0.29.24
decorator==4.4.2
docker-pycreds==0.4.0
filelock==3.3.2
fsspec==2021.11.1
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20211023
gdown==3.13.1
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
gym==0.17.3
huggingface-hub==0.1.2
hydra-core==1.0.5
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-resources==5.4.0
iopath==0.1.9
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.0
joblib==1.1.0
kiwisolver==1.3.2
kornia==0.4.1
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
meshcat==0.0.18
mkdocs-material-extensions==1.0.3
moviepy==1.0.3
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.22.4
oauthlib==3.1.1
olefile==0.46
omegaconf==2.0.6
openai-clip==1.0.1
opencv-python==4.1.2.30
packaging==21.2
pandas==1.3.4
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkbar==0.5
pointnet2==0.0.0
portalocker==2.3.2
proglog==0.1.9
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.0.4
pycocotools==2.0.2
pydot==1.4.2
pyglet==1.5.0
Pygments==2.10.0
pyngrok==5.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.0.3
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-image==0.17.2
scikit-learn==1.0.1
scipy==1.4.1
seaborn==0.11.2
sentencepiece==0.1.97
sentry-sdk==1.4.3
setproctitle==1.3.2
Shapely==1.8.0
shortuuid==1.0.7
six==1.16.0
sklearn==0.0
smmap==5.0.0
subprocess32==3.5.4
tabulate==0.8.9
tdqm==0.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tokenizers==0.10.3
toml==0.10.2
torch-tb-profiler==0.4.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.3.2
transforms3d==0.3.1
typing_extensions==4.3.0
u-msgpack-python==2.7.1
urllib3==1.26.7
wandb==0.13.3
watchdog==0.10.4
wcwidth==0.2.5
Werkzeug==2.0.2
yacs==0.1.8
yaspin==2.1.0
zipp==3.6.0
setuptools==59.5.0
